+++ big_3.c
@@ -6,14 +6,14 @@
 /*   By: dyoula <dyoula@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/24 14:03:53 by dyoula            #+#    #+#             */
-/*   Updated: 2021/12/03 22:06:02 by dyoula           ###   ########.fr       */
+/*   Updated: 2021/12/05 22:08:24 by dyoula           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 #include "includes/push_swap.h"
 
-void	b_which_op_max(int index, t_list *stack)
+void	b_which_op_max(long index, t_list *stack)
 {
 	if (index == 0)
 		ra_list(stack, 'a', stack->l_co);
@@ -23,8 +23,8 @@ void	b_which_op_max(int index, t_list *stack)
 
 void	b_three_arg(t_list *stack_a)
 {
-	int	max;
-	int	index;
+	long	max;
+	long	index;
 
 	max = find_max(stack_a);
 	index = find_index_max(stack_a, max);
